+++ Umpire.h
@@ -1,17 +1,21 @@
 #include "Person.h"
+#include "Person.h"
+#include "Player.h"
+#include "Server.h"
+#include "Receiver.h"
 
 class Umpire : public Person {
     private: 
     // DATA MEMBERS
-        int serverScore;
-        int receiverScore;
+        float serverScore;
+        float receiverScore;
         int serverPoints; 
         int receiverPoints;
 
 
     public:
     // CONSTRUCTORS
-        Umpire(int serverScore, int receiverScore, int serverPoints, int receiverPoints);
+        Umpire(float serverScore, float receiverScore, int serverPoints, int receiverPoints);
         Umpire();
 
     // FUNCTIONS
@@ -20,8 +24,8 @@ class Umpire : public Person {
         void announceWinner(string serverName, string receiverName); 
 
     // METHODS
-        int get_serverScore();
-        void set_serverScore(int serverScore);
-        int get_receiverScore();
-        void set_receiverScore(int receiverScore);
+        float get_serverScore();
+        void set_serverScore(float serverScore);
+        float get_receiverScore();
+        void set_receiverScore(float receiverScore);
 };
